Revert "Feature/auth/update/user main page"
