--- src/styles/UserMainForm.scss
@@ -1,117 +1,0 @@
-#main {
-  background-color: rgb(231, 216, 197);
-  position: absolute;
-  left: 0;
-  right: 0;
-  top: 0;
-  bottom: 0;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  min-height: 710px;
-
-  .main {
-    background-image: url("../../public/img/alabaster-co-DC98wz4vNws-unsplash.jpg");
-    background-size: 50rem 50rem;
-    opacity: 0.9;
-    width: 50rem;
-    padding-top: 1rem;
-    padding-bottom: 4rem;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    border-radius: 4px;
-    .logo {
-      display: flex;
-      padding-left: 0.5rem;
-      margin-top: 3rem;
-      &:first-child {
-        font-size: 3.5rem;
-        padding-bottom: 2.3rem;
-        color: #f2f8ff;
-      }
-      span {
-        padding-top: 0.5rem;
-        letter-spacing: 1px;
-        font-family: "Lexend Deca", sans-serif;
-        font-size: 2.25rem;
-        font-weight: bold;
-        text-shadow: 1px 1px 3px #5f5f5f;
-      }
-    }
-    form {
-      display: flex;
-      padding: 0.5rem 1.5rem;
-      input {
-        border: none;
-        background-color: inherit;
-        font-size: 1.3rem;
-        padding-left: 0.8rem;
-        margin-left: 1rem;
-        font-family: Jua;
-        border-bottom: 0.125rem solid #eef7fc;
-        color: #eef7fc;
-        &::placeholder {
-          color: #eef7fc;
-        }
-        &:focus {
-          outline: none;
-        }
-      }
-      button {
-        display: flex;
-        align-items: center;
-        border: none;
-        background-color: inherit;
-        .icon {
-          color: white;
-          font-size: 1.6rem;
-          border-radius: 50%;
-          padding: 2px;
-          transition: 0.5s;
-          cursor: pointer;
-          &:hover {
-            box-shadow: 1px 1px 3px white;
-          }
-        }
-      }
-    }
-  }
-  .feature-box {
-    margin-top: 2rem;
-    height: 25rem;
-    display: flex;
-    justify-content: center;
-    .feature {
-      cursor: pointer;
-      div {
-        height: 10rem;
-        width: 10rem;
-        margin: 1rem;
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-        justify-content: center;
-        border-radius: 1rem;
-        background-color: #30303077;
-        box-shadow: 1px 1px 3px;
-        transition: 0.7s;
-        &:hover {
-          background-color: #000000de;
-        }
-      }
-      .icon {
-        padding: 1rem;
-        font-size: 2rem;
-        color: #e8f8ff;
-        border: 0.15rem solid #e8f8ffc5;
-        border-radius: 4px;
-      }
-      .text {
-        font-family: Jua;
-        margin-top: 1rem;
-        color: #e8f8ff;
-      }
-    }
-  }
-}
