+++ src/styles/components/UserPropfileForm.scss
@@ -8,7 +8,7 @@
   display: flex;
   justify-content: center;
   align-items: center;
-  min-height: 700px;
+  min-height: 530px;
   .profile {
     width: 360px;
     min-width: 360px;
@@ -18,7 +18,7 @@
     box-shadow: 1px 1px 8px #9fa3a7d0;
     .logo {
       display: flex;
-      padding-left: 1.3rem;
+      padding-left: 0.5rem;
       &:first-child {
         font-size: 3.5rem;
         padding-bottom: 2rem;
@@ -42,40 +42,21 @@
     padding-bottom: 10px;
     div {
       display: flex;
-      align-items: flex-end;
+      align-items: center;
       margin-top: 10px;
       padding: 10px;
       color: #492300;
       opacity: 0.9;
+      .edit-btn {
+        font-size: 1.3rem;
+        opacity: 0.8;
+      }
     }
-    .title {
+    span {
       background-color: #dfdfdf;
       padding: 5px;
       margin-top: 5px;
     }
-    .editBtn {
-      transition: 1s;
-      span {
-        font-size: 1.5rem;
-        font-family: "Nanum Gothic Coding", monospace;
-        margin-left: 0.4rem;
-      }
-      button {
-        display: flex;
-        background-color: inherit;
-        padding: 0.2rem 2rem;
-        font-size: 1.4rem;
-        margin-top: 2rem;
-        border: none;
-        border-radius: 5px;
-        transition: 0.8s;
-        &:hover {
-          background-color: #331a00;
-          color: white;
-        }
-      }
-    }
-
     input {
       margin-left: 0.7rem;
       margin-right: 0.6rem;
